+++ src/components/CardBox.js
@@ -1,17 +1,6 @@
-export default function CardBox({ cardNumber, cardOwnerName, cardExpiration }) {
-  function maskingCardNumber(cardNumber) {
-    const cardNumberArray = Object.values(cardNumber);    
-    return cardNumberArray.map((value, index) => {
-      return (index < 3 && value.length === 4)? value + "-" : value;
-    }).join("");
-  }
+import { maskingCardNumber, maskingCardExpiration } from "../utils/card";
 
-  function maskingCardExpiration(cardExpiration) {
-    const cardExpirationArray = Object.values(cardExpiration);
-    return cardExpirationArray.map((value, index) => {
-      return (index == 0 && value < 13 && value > 0)? value + " / " : value;
-    }).join("");
-  }
+export default function CardBox({ cardNumber, cardOwnerName, cardExpiration }) {  
   return (
     <div className="card-box">
       <div className="empty-card">
